requirements.txt
Django==2.2.5
dj-database-url==0.5.0
psycopg2-binary==2.8.3
djangorestframework==3.10.3

# dev depencies
pytest==4.0.2
pytest-django==3.4.4
model-mommy==1.6.0
pytest-mock==1.10.0
